moved Alignment into player.utils package
